requirements.txt
googletrans
beautifulsoup4
pdfminer.six
